+++ HISTORY.rst
@@ -8,6 +8,7 @@ History
 
 - Dropped previously-deprecated ``Charge.fee_details`` property.
 - Dropped previously-deprecated ``Transfer.fee_details`` property.
+- Dropped previously-deprecated ``field_name`` parameter to ``sync_from_stripe_data``
 
 2.0.3 (2019-06-11)
 ------------------